+++ app/src/main/java/com/kickstarter/features/videofeed/ui/VideoFeedActivity.kt
@@ -4,29 +4,12 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.pager.VerticalPager
-import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
+import androidx.compose.runtime.getValue
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.kickstarter.features.videofeed.data.KSVideoBadgeType
-import com.kickstarter.features.videofeed.ui.components.KSVideoActionsColumn
-import com.kickstarter.features.videofeed.ui.components.KSVideoBadgesRow
-import com.kickstarter.features.videofeed.ui.components.KSVideoCampaignCard
 import com.kickstarter.features.videofeed.viewmodel.VideoFeedViewModel
 import com.kickstarter.libs.Environment
 import com.kickstarter.libs.utils.extensions.getEnvironment
 import com.kickstarter.ui.compose.designsystem.KSTheme
-import com.kickstarter.ui.compose.designsystem.KSTheme.dimensions
-import com.kickstarter.ui.compose.designsystem.videoplayer.KSVideoPlayer
-import kotlin.getValue
 
 class VideoFeedActivity : ComponentActivity() {
 
@@ -44,78 +27,8 @@ class VideoFeedActivity : ComponentActivity() {
 
         setContent {
             KSTheme {
-                // TODO: In future tickets this hardcoded list will be substituted by the result of a query
-                val badges = listOf(
-                    KSVideoBadgeType.ProjectWeLove,
-                    KSVideoBadgeType.DaysLeft("3 days left"),
-                    KSVideoBadgeType.JustLaunched,
-                    KSVideoBadgeType.Trending
-                )
-
-                // TODO: Following vm lines are for qa/demo purposes as of now!!! aiming here to avoid videoURL expiration real VM still to come
-                val videoFeedUIState = viewModel.videoFeedUIState.collectAsStateWithLifecycle()
-                val projectsList = videoFeedUIState.value.projects
-
-                val pagerState = rememberPagerState(pageCount = { projectsList.size })
-
-//                val pool = remember { VideoPlayerPool(this) }
-//                DisposableEffect(pool) { onDispose { pool.releaseAll() } }
-
-                VerticalPager(
-                    modifier = Modifier.fillMaxSize(),
-                    state = pagerState,
-                    beyondViewportPageCount = 1,
-                    key = { index -> projectsList[index].id() }
-                ) { page ->
-
-                    val project = projectsList[page]
-                    val videoUrl = project.video()?.hls() ?: ""
-                    val profileImage = project.creator()?.avatar()?.medium() ?: ""
-                    val projectTitle = project.name()
-                    val percentageFounded = project.percentageFunded()
-
-//                    val player = remember(page, videoUrl) {
-//                        pool.getPlayer(page, videoUrl, pagerState.currentPage)
-//                    }
-
-                    KSVideoPlayer(
-                        videoUrl = videoUrl,
-                        isActive = pagerState.currentPage == page,
-                        overlayContent = { hazeState ->
-                            Column(
-                                modifier = Modifier.fillMaxWidth()
-                            ) {
-                                KSVideoActionsColumn(
-                                    modifier = Modifier
-                                        .align(Alignment.End)
-                                        .padding(end = dimensions.paddingMediumLarge),
-                                    profileImageUrl = profileImage,
-                                    bookmarkCount = "1k",
-                                    shareCount = "50",
-                                    onProfileClick = { },
-                                    onBookmarkClick = { },
-                                    onShareClick = { },
-                                    onMoreOptionsClick = { }
-                                )
-
-                                Spacer(modifier = Modifier.height(24.dp))
-
-                                KSVideoBadgesRow(
-                                    badges = badges,
-                                    hazeState = hazeState
-                                )
-
-                                KSVideoCampaignCard(
-                                    title = projectTitle,
-                                    subtitle = "$50,134 pledged • Join 431 backers",
-                                    buttonText = "Back this project",
-                                    onButtonClick = { },
-                                    progress = percentageFounded
-                                )
-                            }
-                        }
-                    )
-                }
+                val videoFeedUIState by viewModel.videoFeedUIState.collectAsStateWithLifecycle()
+                VideoFeedScreen(projectsList = videoFeedUIState.projects)
             }
         }
     }
